requirements.txt
# Core pipeline
anthropic>=0.42.0
google-api-python-client>=2.150.0
google-auth>=2.35.0
google-auth-oauthlib>=1.2.0

# Cloud Function framework
functions-framework>=3.8.0

# Testing / dev
Pillow>=10.0.0
